fix(mr comments): dispose threads before displaying comments.
